rename rpc_state -> rpc_io
